Subjects statistics missing retention
