+++ src/main/java/br/edu/ufcg/computacao/eureca/backend/core/controllers/SubjectsStatisticsController.java
@@ -1,5 +1,6 @@
 package br.edu.ufcg.computacao.eureca.backend.core.controllers;
 
+import br.edu.ufcg.computacao.eureca.backend.constants.Messages;
 import br.edu.ufcg.computacao.eureca.backend.constants.PortugueseStudentsGlossary;
 import br.edu.ufcg.computacao.eureca.backend.api.http.response.SubjectSummaryResponse;
 import br.edu.ufcg.computacao.eureca.backend.api.http.response.SubjectsSummaryItemResponse;
@@ -7,6 +8,7 @@ import br.edu.ufcg.computacao.eureca.backend.core.dao.DataAccessFacade;
 import br.edu.ufcg.computacao.eureca.backend.core.holders.DataAccessFacadeHolder;
 import br.edu.ufcg.computacao.eureca.backend.core.holders.EnviromentVariablesHolder;
 import br.edu.ufcg.computacao.eureca.backend.core.models.*;
+import br.edu.ufcg.computacao.eureca.common.exceptions.InvalidParameterException;
 import org.apache.log4j.Logger;
 
 import java.util.*;
@@ -29,27 +31,31 @@ public class SubjectsStatisticsController {
                 .collect(Collectors.toCollection(TreeSet::new));
     }
 
-    public SubjectSummaryResponse getSubjectStatistics() {
+    public SubjectSummaryResponse getSubjectStatistics() throws InvalidParameterException {
         String course = EnviromentVariablesHolder.getInstance().getEnvironmentVariables().getCurrentCourse();
         String code = EnviromentVariablesHolder.getInstance().getEnvironmentVariables().getCurrentCurriculum();
-        CurriculumParsedData curriculum = new CurriculumParsedData(this.dataAccessFacade.getCurriculum(course, code));
+        Curriculum curriculum = this.dataAccessFacade.getCurriculum(course, code);
+        if (curriculum == null) {
+            throw new InvalidParameterException(String.format(Messages.INEXISTENT_CURRICULUM_S_S, code, course));
+        }
         SubjectSummaryResponse summary = getSubjectStatisticsSummary(code, curriculum);
         return summary;
     }
 
-    private SubjectSummaryResponse getSubjectStatisticsSummary(String curriculumCode, CurriculumParsedData curriculum) {
-        SubjectStatisticsItem mandatory = buildSummary(curriculumCode, curriculum.getMandatorySubjects());
-        SubjectStatisticsItem optional = buildSummary(curriculumCode, curriculum.getOptionalSubjects());
-        SubjectStatisticsItem elective = buildSummary(curriculumCode, curriculum.getElectiveSubjects());
-        SubjectStatisticsItem complementary = buildSummary(curriculumCode, curriculum.getComplementarySubjects());
-        TreeSet<String> terms = this.dataAccessFacade.getTerms(curriculumCode);
+    private SubjectSummaryResponse getSubjectStatisticsSummary(String curriculumCode, Curriculum curriculum) {
+        SubjectStatisticsItem mandatory = buildSummary(curriculumCode, curriculum.getMandatorySubjectsList());
+        SubjectStatisticsItem optional = buildSummary(curriculumCode, curriculum.getOptionalSubjectsList());
+        SubjectStatisticsItem elective = buildSummary(curriculumCode, curriculum.getElectiveSubjectsList());
+        SubjectStatisticsItem complementary = buildSummary(curriculumCode, curriculum.getComplementarySubjectsList());
+        TreeSet<String> terms = this.dataAccessFacade.getTermsForCurriculum(curriculumCode);
         String from = terms.first();
         String to = terms.last();
-        SubjectSummaryResponse ret = new SubjectSummaryResponse(curriculumCode, from, to, mandatory, optional, elective, complementary);
+        SubjectSummaryResponse ret = new SubjectSummaryResponse(curriculumCode, from, to, mandatory, optional,
+                elective, complementary);
         return ret;
     }
 
-    private SubjectStatisticsItem buildSummary(String curriculumCode, String[] subjects) {
+    private SubjectStatisticsItem buildSummary(String curriculumCode, Collection<String> subjects) {
         int minFailedDueToAbsences = Integer.MAX_VALUE;
         int maxFailedDueToAbsences = 0;
         int totalFailedDueToAbsences = 0;
@@ -62,37 +68,39 @@ public class SubjectsStatisticsController {
         int minSucceeded = Integer.MAX_VALUE;
         int maxSucceeded = 0;
         int totalSucceeded = 0;
+        int totalNumberOfClasses = 0;
 
-        for (int i = 0; i < subjects.length; i++) {
-            SubjectParsedData subject = new SubjectParsedData(subjects[i], this.dataAccessFacade.getSubject(subjects[i]));
-            EnrollmentStatistics failedDueToAbsences = this.dataAccessFacade.getStatisticsFailedDueToAbsences(curriculumCode, subject.getCode());
+        for(String item : subjects) {
+            totalNumberOfClasses += this.dataAccessFacade.getNumberOfClassesPerSubject(curriculumCode, item);
+            Subject subject = this.dataAccessFacade.getSubject(item);
+            EnrollmentStatistics failedDueToAbsences = this.dataAccessFacade.getStatisticsFailedDueToAbsences(curriculumCode, item);
             if (minFailedDueToAbsences > failedDueToAbsences.getMin()) minFailedDueToAbsences = failedDueToAbsences.getMin();
             if (maxFailedDueToAbsences < failedDueToAbsences.getMax()) maxFailedDueToAbsences = failedDueToAbsences.getMax();
             totalFailedDueToAbsences += failedDueToAbsences.getTotal();
-            EnrollmentStatistics failedDueToGrade = this.dataAccessFacade.getStatisticsFailedDueToGrade(curriculumCode, subject.getCode());
+            EnrollmentStatistics failedDueToGrade = this.dataAccessFacade.getStatisticsFailedDueToGrade(curriculumCode, item);
             if (minFailedDueToGrade > failedDueToGrade.getMin()) minFailedDueToGrade = failedDueToGrade.getMin();
             if (maxFailedDueToGrade < failedDueToGrade.getMax()) maxFailedDueToGrade = failedDueToGrade.getMax();
             totalFailedDueToGrade += failedDueToGrade.getTotal();
-            EnrollmentStatistics suspended = this.dataAccessFacade.getStatisticsSuspended(curriculumCode, subject.getCode());
+            EnrollmentStatistics suspended = this.dataAccessFacade.getStatisticsSuspended(curriculumCode, item);
             if (minSuspended > suspended.getMin()) minSuspended = suspended.getMin();
             if (maxSuspended < suspended.getMax()) maxSuspended = suspended.getMax();
             totalSuspended += suspended.getTotal();
-            EnrollmentStatistics succeeded = this.dataAccessFacade.getStatisticsSucceeded(curriculumCode, subject.getCode());
+            EnrollmentStatistics succeeded = this.dataAccessFacade.getStatisticsSucceeded(curriculumCode, item);
             if (minSucceeded > succeeded.getMin()) minSucceeded = succeeded.getMin();
             if (maxSucceeded < succeeded.getMax()) maxSucceeded = succeeded.getMax();
             totalSucceeded += succeeded.getTotal();
         }
         MetricSummary failedDueToAbsencesSummary = new MetricSummary(minFailedDueToAbsences, maxFailedDueToAbsences,
-                (1.0*totalFailedDueToAbsences)/subjects.length);
+                (1.0*totalFailedDueToAbsences)/totalNumberOfClasses);
         MetricSummary failedDueToGradeSummary = new MetricSummary(minFailedDueToGrade, maxFailedDueToGrade,
-                (1.0*totalFailedDueToGrade)/subjects.length);
+                (1.0*totalFailedDueToGrade)/totalNumberOfClasses);
         MetricSummary suspendedSummary = new MetricSummary(minSuspended, maxSuspended,
-                (1.0*totalSuspended)/subjects.length);
+                (1.0*totalSuspended)/totalNumberOfClasses);
         MetricSummary succeededSummary = new MetricSummary(minSucceeded, maxSucceeded,
-                (1.0*totalSucceeded)/subjects.length);
+                (1.0*totalSucceeded)/totalNumberOfClasses);
         MetricSummary relativeRetentionSummary = new MetricSummary(0, 0, 0.0);
         MetricSummary absoluteRetentionSummary = new MetricSummary(0, 0, 0.0);
-        SubjectStatisticsItem ret = new SubjectStatisticsItem(subjects.length, 0, 0,
+        SubjectStatisticsItem ret = new SubjectStatisticsItem(subjects.size(),
                 failedDueToAbsencesSummary, failedDueToGradeSummary, suspendedSummary, succeededSummary,
                 relativeRetentionSummary, absoluteRetentionSummary);
         return ret;
@@ -100,10 +108,10 @@ public class SubjectsStatisticsController {
 
     public SubjectSummaryResponse getSubjectStatisticsMock() {
         MetricSummary metrics = new MetricSummary(20, 30, 25);
-        SubjectStatisticsItem mandatory = new SubjectStatisticsItem(1980, 196, 30,metrics,metrics, metrics, metrics, metrics, metrics);
-        SubjectStatisticsItem optional = new SubjectStatisticsItem(400, 50, 10, metrics, metrics, metrics, metrics, metrics, metrics);
-        SubjectStatisticsItem elective = new SubjectStatisticsItem(840, 56,15,metrics, metrics, metrics, metrics, metrics, metrics);
-        SubjectStatisticsItem complementary = new SubjectStatisticsItem(450, 30, 10, metrics, metrics, metrics, metrics, metrics, metrics);
+        SubjectStatisticsItem mandatory = new SubjectStatisticsItem(30,metrics,metrics, metrics, metrics, metrics, metrics);
+        SubjectStatisticsItem optional = new SubjectStatisticsItem(10, metrics, metrics, metrics, metrics, metrics, metrics);
+        SubjectStatisticsItem elective = new SubjectStatisticsItem(15,metrics, metrics, metrics, metrics, metrics, metrics);
+        SubjectStatisticsItem complementary = new SubjectStatisticsItem(10, metrics, metrics, metrics, metrics, metrics, metrics);
         SubjectSummaryResponse summary = new SubjectSummaryResponse("2017", "1980.1", "2020.1", mandatory, optional, elective, complementary);
         return summary;
     }
