+++ src/main/java/br/edu/ufcg/computacao/eureca/backend/core/dao/scsvfiles/mapentries/AcademicUnitData.java
@@ -1,4 +1,20 @@
 package br.edu.ufcg.computacao.eureca.backend.core.dao.scsvfiles.mapentries;
 
-public class AcademicUnitData extends br.edu.ufcg.computacao.eureca.backend.core.models.AcademicUnitData implements EurecaMapValue {
+public class AcademicUnitData implements EurecaMapValue {
+    private String acronym;
+
+    public String getAcronym() {
+        return acronym;
+    }
+
+    public AcademicUnitData() {
+    }
+
+    public void setAcronym(String acronym) {
+        this.acronym = acronym;
+    }
+
+    public AcademicUnitData(String acronym) {
+        this.acronym = acronym;
+    }
 }
