+++ src/main/java/br/edu/ufcg/computacao/eureca/backend/core/dao/scsvfiles/mapentries/AcademicUnitKey.java
@@ -1,4 +1,35 @@
 package br.edu.ufcg.computacao.eureca.backend.core.dao.scsvfiles.mapentries;
 
-public class AcademicUnitKey extends br.edu.ufcg.computacao.eureca.backend.core.models.AcademicUnitKey implements EurecaMapKey {
+import java.util.Objects;
+
+public class AcademicUnitKey implements EurecaMapKey {
+    private String code;
+
+    public AcademicUnitKey(String code) {
+        this.code = code;
+    }
+
+    public AcademicUnitKey() {
+    }
+
+    public String getCode() {
+        return code;
+    }
+
+    public void setCode(String code) {
+        this.code = code;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+        AcademicUnitKey that = (AcademicUnitKey) o;
+        return getCode().equals(that.getCode());
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(getCode());
+    }
 }
