--- src/main/java/br/edu/ufcg/computacao/eureca/backend/core/dao/scsvfiles/mapentries/CpfRegistrationKey.java
@@ -1,4 +1,0 @@
-package br.edu.ufcg.computacao.eureca.backend.core.dao.scsvfiles.mapentries;
-
-public class CpfRegistrationKey extends br.edu.ufcg.computacao.eureca.backend.core.models.CpfRegistrationKey implements EurecaMapKey {
-}
