+++ src/main/java/br/edu/ufcg/computacao/eureca/backend/core/dao/scsvfiles/mapentries/SubjectKey.java
@@ -1,4 +1,35 @@
 package br.edu.ufcg.computacao.eureca.backend.core.dao.scsvfiles.mapentries;
 
-public class SubjectKey extends br.edu.ufcg.computacao.eureca.backend.core.models.SubjectKey implements EurecaMapKey {
+import java.util.Objects;
+
+public class SubjectKey implements EurecaMapKey {
+    private String code;
+
+    public SubjectKey(String code) {
+        this.code = code;
+    }
+
+    public SubjectKey() {
+    }
+
+    public String getCode() {
+        return code;
+    }
+
+    public void setCode(String code) {
+        this.code = code;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+        SubjectKey that = (SubjectKey) o;
+        return getCode().equals(that.getCode());
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(getCode());
+    }
 }
