+++ src/main/java/br/edu/ufcg/computacao/eureca/backend/core/dao/scsvfiles/mapentries/TeacherData.java
@@ -1,4 +1,20 @@
 package br.edu.ufcg.computacao.eureca.backend.core.dao.scsvfiles.mapentries;
 
-public class TeacherData extends br.edu.ufcg.computacao.eureca.backend.core.models.TeacherData implements EurecaMapValue {
+public class TeacherData implements EurecaMapValue {
+    private String teachers;
+
+    public TeacherData(String teachers) {
+        this.teachers = teachers;
+    }
+
+    public TeacherData() {
+    }
+
+    public String getTeachers() {
+        return teachers;
+    }
+
+    public void setTeachers(String teachers) {
+        this.teachers = teachers;
+    }
 }
