--- src/main/java/br/edu/ufcg/computacao/eureca/backend/core/models/AcademicUnitData.java
@@ -1,20 +1,0 @@
-package br.edu.ufcg.computacao.eureca.backend.core.models;
-
-public class AcademicUnitData {
-    private String acronym;
-
-    public String getAcronym() {
-        return acronym;
-    }
-
-    public AcademicUnitData() {
-    }
-
-    public void setAcronym(String acronym) {
-        this.acronym = acronym;
-    }
-
-    public AcademicUnitData(String acronym) {
-        this.acronym = acronym;
-    }
-}
