+++ src/test/java/br/edu/ufcg/computacao/eureca/backend/api/http/request/SubjectStatisticsTest.java
@@ -30,10 +30,10 @@ public class SubjectStatisticsTest extends EndpointTest {
 
     private SubjectSummaryResponse getSubjectStatisticsMock() {
         MetricSummary metrics = new MetricSummary(20, 30, 25);
-        SubjectStatisticsItem mandatory = new SubjectStatisticsItem(1980, 196, 30,metrics,metrics, metrics, metrics, metrics, metrics);
-        SubjectStatisticsItem optional = new SubjectStatisticsItem(400, 50, 10, metrics, metrics, metrics, metrics, metrics, metrics);
-        SubjectStatisticsItem elective = new SubjectStatisticsItem(840, 56,15,metrics, metrics, metrics, metrics, metrics, metrics);
-        SubjectStatisticsItem complementary = new SubjectStatisticsItem(450, 30, 10, metrics, metrics, metrics, metrics, metrics, metrics);
+        SubjectStatisticsItem mandatory = new SubjectStatisticsItem(30,metrics,metrics, metrics, metrics, metrics, metrics);
+        SubjectStatisticsItem optional = new SubjectStatisticsItem(10, metrics, metrics, metrics, metrics, metrics, metrics);
+        SubjectStatisticsItem elective = new SubjectStatisticsItem(15,metrics, metrics, metrics, metrics, metrics, metrics);
+        SubjectStatisticsItem complementary = new SubjectStatisticsItem(10, metrics, metrics, metrics, metrics, metrics, metrics);
         SubjectSummaryResponse summary = new SubjectSummaryResponse("2017", "1980.1", "2020.1", mandatory, optional, elective, complementary);
         return summary;
     }
