+++ src/test/java/br/edu/ufcg/computacao/eureca/backend/core/tests/models/NewStudentTest.java
@@ -1,11 +1,11 @@
 package br.edu.ufcg.computacao.eureca.backend.core.tests.models;
 
-import br.edu.ufcg.computacao.eureca.backend.core.models.CpfRegistrationKey;
+import br.edu.ufcg.computacao.eureca.backend.core.dao.scsvfiles.mapentries.NationalIdRegistrationKey;
+import br.edu.ufcg.computacao.eureca.backend.core.dao.scsvfiles.mapentries.StudentData;
 import br.edu.ufcg.computacao.eureca.backend.core.holders.DataAccessFacadeHolder;
-import br.edu.ufcg.computacao.eureca.backend.core.models.RiskClass;
 import br.edu.ufcg.computacao.eureca.backend.core.models.Student;
+import br.edu.ufcg.computacao.eureca.backend.core.models.RiskClass;
 import br.edu.ufcg.computacao.eureca.backend.core.util.MetricsCalculator;
-import br.edu.ufcg.computacao.eureca.backend.core.models.StudentData;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
@@ -17,7 +17,7 @@ import static br.edu.ufcg.computacao.eureca.backend.core.models.RiskClass.*;
 
 @RunWith(PowerMockRunner.class)
 @PrepareForTest({MetricsCalculator.class, DataAccessFacadeHolder.class})
-public class StudentTest {
+public class NewStudentTest {
 
     private Student student;
 
@@ -33,10 +33,10 @@ public class StudentTest {
     @Test
     public void getIdTest() {
         // set up
-        CpfRegistrationKey expected = new CpfRegistrationKey("nationalID","registration1");
+        NationalIdRegistrationKey expected = new NationalIdRegistrationKey("nationalID","registration1");
 
         // exercise
-        CpfRegistrationKey result = this.student.getId();
+        NationalIdRegistrationKey result = new NationalIdRegistrationKey(this.student);
 
         // verify
         Assert.assertEquals(expected, result);
@@ -55,7 +55,7 @@ public class StudentTest {
                 0,0,0);
 
         // exercise
-        StudentData result = this.student.getStudentData();
+        StudentData result = new StudentData(this.student);
 
         // verify
         Assert.assertEquals(expected.getName(), result.getName());
@@ -99,16 +99,16 @@ public class StudentTest {
     }
 
     private Student createNewStudent(String nationalId, String registration, int mandatoryHours, int mandatoryCredits, int electiveHours, int electiveCredits,
-                                  int complementaryHours, int complementaryCredits, double gpa, double mc,
-                                  double iea, int completedTerms, int suspendedTerms, int institutionalTerms,
-                                  int mobilityTerms, int enrolledCredits, double admissionGrade) {
+                                     int complementaryHours, int complementaryCredits, double gpa, double mc,
+                                     double iea, int completedTerms, int suspendedTerms, int institutionalTerms,
+                                     int mobilityTerms, int enrolledCredits, double admissionGrade) {
 
-        CpfRegistrationKey cpfRegistration = new CpfRegistrationKey(nationalId,registration);
-        StudentData studentData = new StudentData("x", "x", "x", "x", "x",
+        NationalIdRegistrationKey cpfRegistration = new NationalIdRegistrationKey(nationalId,registration);
+        br.edu.ufcg.computacao.eureca.backend.core.dao.scsvfiles.mapentries.StudentData studentData = new StudentData("x", "x", "x", "x", "x",
                 "x", "x", "x", "Ativo",
                 "VESTIBULAR 2007.2", "x", "x", "x",
                 "x", mandatoryHours, mandatoryCredits, electiveHours, electiveCredits, complementaryHours, complementaryCredits,
                 gpa, mc, iea, completedTerms, suspendedTerms, institutionalTerms, mobilityTerms, enrolledCredits, admissionGrade);
-        return new Student(cpfRegistration, studentData);
+        return studentData.createStudent(cpfRegistration);
     }
 }
