+++ src/test/java/br/edu/ufcg/computacao/eureca/backend/util/TestUtils.java
@@ -2,8 +2,9 @@ package br.edu.ufcg.computacao.eureca.backend.util;
 
 import br.edu.ufcg.computacao.eureca.backend.api.http.CommonKeys;
 import br.edu.ufcg.computacao.eureca.backend.api.http.response.*;
+import br.edu.ufcg.computacao.eureca.backend.core.dao.scsvfiles.mapentries.NationalIdRegistrationKey;
+import br.edu.ufcg.computacao.eureca.backend.core.dao.scsvfiles.mapentries.StudentData;
 import br.edu.ufcg.computacao.eureca.backend.core.models.Metrics;
-import br.edu.ufcg.computacao.eureca.backend.core.models.StudentData;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpMethod;
 import org.springframework.http.MediaType;
@@ -55,7 +56,8 @@ public class TestUtils {
                 58,0,26,5.68,
                 7,1.69,14,1,0,
                 0,0,0);
-        StudentDataResponse mockedStudentDataResponse = new StudentDataResponse("fake-registration", mockedStudentData);
+        StudentDataResponse mockedStudentDataResponse = new StudentDataResponse(mockedStudentData.
+                createStudent(new NationalIdRegistrationKey("fake-national-id", "fake-registration")));
         return Arrays.asList(mockedStudentDataResponse);
     }
